mv reduce funcs into reduce_cuda_impl
